chore: move utils files
