+++ lua/fugit2/util/stat.lua
