+++ lua/fugit2/view/components/commit_log_view.lua
@@ -7,7 +7,7 @@ local Object = require "nui.object"
 local string_utils = require "plenary.strings"
 
 local git2 = require "fugit2.core.git2"
-local pendulum = require "fugit2.core.pendulum"
+local pendulum = require "fugit2.util.pendulum"
 local utils = require "fugit2.utils"
 
 local TAG_PRE_WIDTH = 40
